pre-commit fixes
